imglib2-scripting: Histogram.computeBin returns a long instead of an int. Given that positions along a dimension are in long, makes more sense.

+++ imglib2/scripting/src/main/java/net/imglib2/script/algorithm/integral/histogram/Histogram.java
@@ -54,7 +54,7 @@ public abstract class Histogram<T extends RealType<T>>
 		this.accessBinValues = this.binValues.randomAccess();
 	}
 	
-	public abstract int computeBin(final T value);
+	public abstract long computeBin(final T value);
 	
 	public T binValue(final long index) {
 		this.accessBinValues.setPosition(index, 0);
